backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.1
django-filter==23.3
python-decouple==3.8
dj-database-url==2.1.0
psycopg2-binary==2.9.7
channels==4.0.0
channels-redis==4.1.0
redis==5.0.1
celery==5.3.2
whitenoise==6.6.0
Pillow==10.0.1
python-json-logger==2.0.7
scikit-learn==1.3.0
pandas==2.0.3
numpy==1.24.3
matplotlib==3.7.2
seaborn==0.12.2
plotly==5.15.0
requests==2.31.0
beautifulsoup4==4.12.2
lxml==4.9.3
openpyxl==3.1.2
PyPDF2==3.0.1
python-docx==0.8.11
faker==19.6.2
factory-boy==3.3.0
pytest==7.4.2
pytest-django==4.5.2
coverage==7.3.2
